requirements.txt
langchain>=0.1.0
langgraph>=0.0.50
openai>=1.0.0
duckduckgo-search>=3.9.0
python-dotenv>=1.0.0
pydantic>=2.0.0